+++ projects/Mallard/src/App.tsx
@@ -27,7 +27,7 @@ import { ErrorBoundary } from './components/layout/ui/errors/error-boundary'
 import { Modal, ModalRenderer } from './components/modal'
 import { NetInfoAutoToast } from './components/toast/net-info-auto-toast'
 import { nestProviders } from './helpers/provider'
-import { pushNotifcationRegistration } from './helpers/push-notifications'
+import { pushNotifcationRegistration } from './push-notifications/push-notifications'
 import { ToastProvider } from './hooks/use-toast'
 import { DeprecateVersionModal } from './screens/deprecate-screen'
 import { errorService } from './services/errors'
