+++ projects/Mallard/src/helpers/clear-download-issue.ts
@@ -1,7 +1,7 @@
 import ApolloClient from 'apollo-client'
 import { fetchCacheClear } from './fetch'
 import { clearOldIssues, downloadTodaysIssue, prepFileSystem } from './files'
-import { cleanPushTrackingByDays } from './push-tracking'
+import { cleanPushTrackingByDays } from '../push-notifications/push-tracking'
 import { largeDeviceMemory } from 'src/hooks/use-config-provider'
 
 const clearAndDownloadIssue = async (client: ApolloClient<object>) => {
