+++ projects/Mallard/src/helpers/files.ts
@@ -10,7 +10,7 @@ import { getSetting } from './settings'
 import { defaultSettings } from './settings/defaults'
 import { errorService } from 'src/services/errors'
 import { londonTime } from './date'
-import { pushTracking } from 'src/helpers/push-tracking'
+import { pushTracking } from 'src/push-notifications/push-tracking'
 import { localIssueListStore } from 'src/hooks/use-issue-on-device'
 import { NetInfo, DownloadBlockedStatus } from 'src/hooks/use-net-info'
 import gql from 'graphql-tag'
